scripts/gen_config.py: add support for hashes
Using a specific hash instead of a branch name for a feed can be
useful when one tries to reproduce an image at two different points in
time.

Allow each feed to have either a "branch" key or a "hash" key, but not
both.

If a hash is specified, the corresponding feed token is used ('^').
Else, the previous behavior remains (using ';').

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>