profiles: move to prplmesh feed-prpl fork
Since we are not allowed to push to prplwrt upstream, shift to use the
feed-prpl fork.
While at it, enable prplmesh in the prpl profile - instead of doing so
in the intel_mips/axepoint profile - since prplmesh is one of the prpl
packages. Select both dwpal and nl80211 variants since the correct
variant will be selected depending on the platform - intel_mips -->
dwpal, others --> nl80211.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>