ci: sdk: provide full logs in case of failure
The first parallel build is quiet and in case of error we build second
time with increased verbosity, but not from scratch. This for example
means, that we're missing output of configure step in the logs which
might be useful in some cases. So now we clean the build dir before
compiling the package again for second time.

Fixes: 8a2a972ec576 ("ci: add support for building of feeds using SDK")
Reported-by: Arnout Vandecappelle <arnout@mind.be>
Signed-off-by: Petr Štetiar <ynezz@true.cz>