ci: sdk: fix package configuration and dependencies
Changes in commit d52303d6c40c ("ci: sdk: speedup builds") deselected
all packages from build config. This lead to the breakage of
feed-prpl/owsd package as it has following config options:

 prompt "Select IPC Bus"
 default OWSD_USE_UBUS
 depends on PACKAGE_owsd

Which means, that it would default to ubus bus if the owsd package is
selected in the config, but this is not the case anymore, so the
configuration step fails:

 CMake Error at CMakeLists.txt:25 (message):
   select at least one of DBus or ubus ...

Fix it by selecting the package in the build config during compile
testing.

Fixes: d52303d6c40c ("ci: sdk: speedup builds")
Signed-off-by: Petr Štetiar <ynezz@true.cz>