netifd: backport: fix IPv6 routing loop on point-to-point links
This is backport of upstream commit c00c8335d618 ("interface-ip: add
unreachable route if address is offlink").

In order to avoid a routing loop add an unreachable route for the
address prefix is the offlink flag is set for an address.  This fixes a
routing loop which is currently present on point-to-point links (e.g
PPP) when the wan interface is assigned a globally unique prefix (e.g.
2001:db8:1:0::/64) from which an IPv6 address is picked and installed on
the wan interface (e.g. 2001:db8:1:0:5054:ff:feab:d87c/64)

The prefix route 2001:db8:1::/64 would be present in the routing table
which will route any packet with as destination 2001:db8:1::/64 to the
wan interface and would be routed back by the upstream router due to the
wan interface due to the assigned global unique prefix.  Besides not
installing the prefix route 2001:db8:1::/64 on point-to-point links
adding an unreachable route is required to avoid the routing loop.

Signed-off-by: Petr Štetiar <ynezz@true.cz>