ci: add prplmesh package to images
In commit 6e1b1e7a73d9 ("ci: rax40: fix failing wifi tests by removing
prplmesh package") prplmesh package (version 1.4.0) was removed because
it was somehow causing CI failures during runtime testing on Netgear
RAX40 target. Meanwhile prplmesh was bumped to version 1.5.0rc6 which
should improve the situation so enable the testing again.

I've removed dhcp_event object from the generic/ubus.t test as it's only
available on MxL targets and should be removed in the future anyway.

That nec-wx3000hp/wifi.t change is caused by the prplmesh as wireless is
enabled by default on this target, so prplmesh starts and configures the
VAPs in the following way, so this is desired state for now.

References: PCF-38
Signed-off-by: Petr Štetiar <ynezz@true.cz>