ci: testbed: glinet-b1300: fix booting of bigger initramfs images
Currently it's not possible to boot 15MB initramfs image as kernel is
not able to decompress itself and is stuck at `Uncompressing Kernel
Image ...` so fix this by lowering the image loading address. Then we
need to adjust default 30 second timeout as it takes roughly 42 seconds
to load initramfs image from network, decompress and boot it.

Signed-off-by: Petr Štetiar <ynezz@true.cz>