profiles: debug: add wpa-cli, iw-full and hostapd debug messages
wpa-cli and iw-full can be useful when debugging on the device itself.

CONFIG_WPA_MSG_MIN_PRIORITY defaults to 3 which makes it impossible to
have hostapd's debug logs. Set it to 0 to allow any log level to be
used.

Note that per [1], this will increase the size of the hostapd
binaries (for debugging purposes, it should be acceptable).

[1]: https://dev.archive.openwrt.org/ticket/8847

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>