profiles: feed-prpl: bump to 2021-05-19 version
1587944 luci-mod-simple: add prplMesh
971d503 prplmesh: uci-defaults: set agent as the default mode
6886438 ci: remove building prplmesh with prplwrt ath79-generic SDK
e210e16 ci: set designated branch to prplwrt-19.07
15d994e ci: allow reuse of pipelines between branches
a10117b ambiorix: add libsahtrace and mod-sahtrace

With this bump agent mode is the default mode for prplMesh. In agent
mode we cannot test that prplMesh becomes operational (because the
agent needs an external controller) while in controller mode we can
check that both the agent and controller are operational. As a
consequence, the tests are adjusted to restart prplMesh in controller
mode by using the prplMesh init script in place of prplmesh_utils.sh.

With the luci-mod-simple modifications there is now a new 'prplmesh'
ubus object, so the ubus cram test is also updated.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
Signed-off-by: Petr Štetiar <ynezz@true.cz>
(cherry picked from commit cdb9554ee3d85878aa670e6cf94594ace6d5151f)