ci: allow reuse of pipelines between branches
The 1.1 release is around the corner, which means, that the maintenance
of that 1.y release would happen in prplwrt-19.07 maintenance branch.

Prepare CI pipelines for easier reusability in other branches by using
`CI_DESIGNATED_BRANCH` variable instead of hardcoded value. This value
specifies branch for which is the CI pipeline designated.

Signed-off-by: Petr Štetiar <ynezz@true.cz>