profiles: feed-prpl: set tracking branch to prplwrt
Set `tracking_branch` to `prplwrt` in order to be consistent with branch
names across repositories wherever possible. Development is going to
happen in `prplwrt` branch and next stable/maintenance branch is going
to be named `prplwrt-19.07` as it would be done in prplwrt repository.

This is purely cosmetic change, nothing currently relies on `tracking_branch`.

Signed-off-by: Petr Štetiar <ynezz@true.cz>