profiles: set feed-prpl's tracking branch back to prplwrt
In `cdb9554ee3d85878aa670e6cf94594ace6d5151f`, the commit hash of
feed_prpl was mistakenly set to the latest commit of
`prplwrt-19.07`. In addition, the `tracking_branch` (which is an
informative field) was set to `prpl-19.07`, reverting the change that
was done in `e5ca8ef5fab0280c4d343ee3f21e0b4e759d9eb4`.

Fix it by setting the tracking branch back to `prplwrt`, and the
commit hash to the latest one of the `prplwrt` branch in feed-prpl.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>