profiles: feed-prpl: add rpcd-mod-luci as diffconfig option
DHCP handling is done with luci-rpc package within PrplMesh.
This package wraps all dhcp leases from both dnsmasq and odhcp as
ubus objects.

PrplWRT based devices should contain this package to serve DHCP
handling. DHCP information for stations are read and registered to
database objects and NBAPI datamodels.

If this package is not enabled, prplMesh works without any problem.
Only, error logs can be seen as "Failed to look up luci-rpc".
This package is not must but helpful in terms of application.
Thus, it is not added as direct dependency.

https://jira.prplfoundation.org/browse/PPM-534
--
Signed-off-by: Cihangir Odabas <codabas@assia-inc.com>