ubus: backport SOVERSION support
Add a support for setting of new `ABIVERSION` CMake define which allows
to control the SOVERSION used for the built shared library. This is
needed for downstream packaging to properly track breaking ABI changes
when updating to newer versions of the library.

Signed-off-by: Felix Fietkau <nbd@nbd.name>
Signed-off-by: Petr Štetiar <ynezz@true.cz> [backport]
(backported from commit 8edb1797d55d259c6eda18c89784f152328436fc)
Signed-off-by: Petr Štetiar <ynezz@true.cz>