ci: lower amount of jobs when scheduled
As job rules are evaluated in order until the first match, currently
when scheduled pipeline is triggered, all jobs will be run as they match
the branch rule which is defined first.  This is not optimal as it
results in a run of all 22 jobs in all 8 stages, which is waste of
resources. Fix that by reordering the job rules, so the pipeline source
trigger is evaluated first, so the jobs are not considered when
triggered by schedule.

Signed-off-by: Petr Štetiar <ynezz@true.cz>