package: hostapd: remove superfluous patch
This patch was integrated in prplOS after being submitted, but not yet
approved (see [1]), to upstream hostapd to fix an issue with backhaul
BSS roaming.

In the end, the proposed patch was superseded and the same issue was
fixed in hostapd with commit
`58bbbb5981440da508164cbd423ca5dd7c1b98d8` (see [2]).

The hostapd commit was backported in prplOS with commit
`23a9f0b255b19e941db3148ce0ac88535980ceb3`, but the initial patch was
not removed. Since it's redundant, remove it now.

Note that keeping both patches should be harmless, so this change
doesn't necessarily have to be backported to prplOS' release branches.

[1]: http://patchwork.ozlabs.org/project/hostap/patch/20210208172823.11536-1-raphael.melotte@mind.be/
[2]: https://w1.fi/cgit/hostap/commit/?id=58bbbb5981440da508164cbd423ca5dd7c1b98d8

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>