PROVISORY: package: dnsmasq: expose more lease data over ubus
For the complete functionality of ambiorix dhcpv4-manager plugin, we
would need this patch included in prplOS. This patch exposes the
requested options of the client, expands the information in the
broadcasted events in ubus, and introduces a method in dnsmasq ubus
object called leases().

* Add dnsmasq networkid to the events.
* Expose DHCPREQUEST options.
* 'leases' ubus method to retrieve the list of leases.

The purpose of the patch is to be able to synchronize with leases
from dnsmasq using ubus only. The ubus 'leases' method can be called
once and then subscribe to the events based on the dhcp packets
received.

The ubus events remain as they were, just add the parameter of the
networkid to identify the pool to which they belong in the config,
and the requested dhcp options from the client.

References: MR!144
References: PCF-557
Signed-off-by: Eduardo Aguilar <eduardo.aguilar_ext@softathome.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz> [commit description]
Signed-off-by: Petr Štetiar <ynezz@true.cz>