ci: single click pipeline run on merge requests
Currently it's necessary to manually click several times on CI's stage
play button in order to execute various stages manually and thus pass
the CI pipeline which is cumbersome. So remove manual rules in two
stages and thus reduce the amount of manual actions to a single click on
the build stage which is the most expensive one.

Signed-off-by: Petr Štetiar <ynezz@true.cz>