scripts: gen_config.py: allow mixing of packages from multiple feeds
Currently there is only toplevel definition of `packages` which should
be installed and we always install all packages from the specified
feeds. This doesn't work well if we would like to mix for example
packages from 19.07 and 21.02 release branches of OpenWrt packages feed.
So this commit adds support for exactly such use case:

  - name: packages_1907
    uri: https://gitlab.com/openwrt/feed/packages.git
    tracking_branch: openwrt-19.07
    revision: 70bb6f15e8dfe5fa2ed5c6d670e661c03949b133
    packages:
      - chrony

  - name: packages_2102
    uri: https://gitlab.com/openwrt/feed/packages.git
    tracking_branch: openwrt-21.02
    revision: b1d69822db7ac61a536f900d0a30d73f714f85a6
    packages:
      - libgpg-error
      - libgpgme
      - libassuan

Signed-off-by: Petr Štetiar <ynezz@true.cz>