scripts: gen_config.py: error out on same feed revisions conflicts
We can currently generate config from multiple profiles:

 ./scripts/gen_config.py ipq40xx prpl webui

If prpl.yml and webui.yml profiles reference same prpl feed with
different revisions, then the resulting config would use revision
defined in the webui.yml profile as it's specified and parsed as latest.

This is most likely unwanted and should be handled as error as the
revisions for the same feed name/url should match, so let's handle this
corner case as error from now on.

Signed-off-by: Petr Štetiar <ynezz@true.cz>