scripts: gen_config.py: allow disabling of default feeds
Currently we install and process all default feeds specified in
feeds.conf.default file which might not be necessary in some use cases,
where for example we've definition of all feeds in the profile YAML
files. In such cases it makes no sense to waste resources on feeds which
wouldn't be used later anyway, so add a boolean profile directive
`use_default_feeds` which would allow disabling of those default feeds
when set to False (default is True).

Signed-off-by: Petr Štetiar <ynezz@true.cz>