ci: tests: cram: generic: ubus: fix Time.CurrentLocalTime test
Current test is racy as we compare timestamp returned by
Time.CurrentLocalTime with timestamp returned by date command formatted
with %Y-%m-%dT%H:%M which means, that if we run the test in 59 second,
then we might get different results as the minute in the command results
wouldn't match.

So lets just convert timestamps to Unix epoch and check if the
difference is within 5s tolerance. While at it add also syslog debug
output so we've some clue if the test fails next time:

 Time.CurrentLocalTime=Fri Jan 28 13:07:27 2022 SystemTime=Fri Jan 28 13:07:27 2022 diff=0s tolerance=5s

Fixes: PCF-543
Suggested-by: Vincent Harlé <vincent.harle@softathome.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz>